requirements.txt
pillow
numpy
tqdm
